Introduce `AppContext`
Rather than the Api and Web classes extending App, we are transitioning
to a 'App has a context' relationship.  We need a `Worker` context and
some of the `App` logic (HTTP Request/Response handling) is not directly
relevant in such a context.  By transitioning to a 'has a' relationship,
the Api and Web contexts can re-use code without forcing the Worker to
also share the HTTP logic.