Use a separate DB connection for worker job management
Jobs need to be able to control their own database transaction.  The
query to determine the next job uses `SELECT ... FOR UPDATE`, which also
requires control of its own transaction.  As such, we need two separate
transactions.  By having the worker use an explicit database connection,
we leave `db()` to be used by the job logic.