Fix asset buster looking at modified time on diff machines
The modified time for our assets are different on each computer, so we
would end up getting a modified manifest file every time we load a page
on a different computer.

I updated it to rely on the sha1 of the file instead of the modified
time for now. I think this could be a performance issue on dev in the
future if we have a lot of assets, and we could potentially update it to
store a separate file for the modified times on dev. But for now I think
this is the most simple solution.